cleaning up unwanted code
