--- AngularApp/projects/app-service-diagnostics/src/app/shared/components/number-picker/number-picker.component.html
@@ -1,14 +1,0 @@
-<div>
-  <div style="display: table;">
-    <div style="display: table-row">
-      <div style="display: table-cell;" class="input-stepper">
-        <button (click)="decrement()" class="minus" [disabled]="value === min"><i class="fa fa-minus"></i></button>
-        <input type="text" [value]="value" disabled="true" />
-        <button (click)="increment()" class="plus" [disabled]="value === max"><i class="fa fa-plus"></i></button>
-      </div>
-      <div style="display: table-cell;">
-        <span class="input-label">{{ additionalText }}</span>
-      </div>
-    </div>
-  </div>
-</div>
