--- AngularApp/projects/app-service-diagnostics/src/app/shared/components/number-picker/number-picker.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { NumberPickerComponent } from './number-picker.component';
-
-describe('NumberPickerComponent', () => {
-  let component: NumberPickerComponent;
-  let fixture: ComponentFixture<NumberPickerComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ NumberPickerComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(NumberPickerComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
